# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Steel Fabrication MRP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      description: List all jobs, optionally filtered by status.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, on_hold, complete, closed]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: A list of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      description: Create a job with an optional custom stage routing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job with full detail
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /jobs/{jobId}/advance:
    post:
      operationId: advanceJobStage
      tags: [jobs]
      summary: Advance job to the next stage
      description: Marks the current in-progress stage complete and starts the next stage.
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Updated job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /time-clock/out:
    post:
      operationId: clockOut
      tags: [time]
      summary: Clock out
      description: Shop floor clock out for an open time entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutInput"
      responses:
        "200":
          description: Clocked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Job:
      type: object
      required:
        - id
        - jobNumber
        - name
        - customer
        - status
        - createdAt
        - updatedAt
        - currentStageName
        - estimatedHours
        - actualHours
        - percentComplete
        - isPastDue
      properties:
        id:
          type: integer
        jobNumber:
          type: string
        name:
          type: string
        customer:
          type: string
        customerId:
          type: ["integer", "null"]
        customerPo:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, on_hold, complete, closed]
        dueDate:
          type: ["string", "null"]
          description: Calendar date (YYYY-MM-DD)
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        currentStageName:
          type: ["string", "null"]
        estimatedHours:
          type: number
        actualHours:
          type: number
        percentComplete:
          type: number
        isPastDue:
          type: boolean
        estimateId:
          type: ["integer", "null"]
        bidNumber:
          type: ["string", "null"]
    JobDetail:
      type: object
      required:
        - id
        - jobNumber
        - name
        - customer
        - status
        - createdAt
        - updatedAt
        - stages
        - estimatedHours
        - actualHours
        - percentComplete
        - isPastDue
        - assignedEmployees
      properties:
        id:
          type: integer
        jobNumber:
          type: string
        name:
          type: string
        customer:
          type: string
        customerId:
          type: ["integer", "null"]
        customerPo:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, on_hold, complete, closed]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        currentStageName:
          type: ["string", "null"]
        estimatedHours:
          type: number
        actualHours:
          type: number
        percentComplete:
          type: number
        isPastDue:
          type: boolean
        estimateId:
          type: ["integer", "null"]
        bidNumber:
          type: ["string", "null"]
        stages:
          type: array
          items:
            $ref: "#/components/schemas/Stage"
        assignedEmployees:
          type: array
          items:
            $ref: "#/components/schemas/AssignedEmployee"
        bomAssemblies:
          type: array
          items:
            $ref: "#/components/schemas/BomAssembly"
        assemblyCount:
          type: integer
        assemblyProgressPct:
          type: [number, "null"]
        assemblyStatus:
          type: [string, "null"]
    JobStageInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        estimatedHours:
          type: number
...
    JobInput:
      type: object
      required: [name, customerId]
      properties:
        name:
          type: string
          minLength: 1
        customerId:
          type: integer
        status:
          type: string
          enum: [active, on_hold, complete, closed]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        assignedEmployeeIds:
          type: array
          description: Employee ids to assign to the job.
          items:
            type: integer
        stages:
          type: array
          description: Custom routing. If omitted, the default workflow is applied.
          items:
            $ref: "#/components/schemas/JobStageInput"
    JobUpdate:
      type: object
      properties:
        jobNumber:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        customerId:
          type: integer
        customerPo:
          type: ["string", "null"]
          description: Customer purchase order number. Null clears it.
        status:
          type: string
          enum: [active, on_hold, complete, closed]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        assignedEmployeeIds:
          type: array
          description: Full replacement list of assigned employee ids.
          items:
            type: integer
    AssignedEmployee:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    Stage:
      type: object
      required:
        - id
        - jobId
        - name
        - orderIndex
        - status
        - estimatedHours
        - actualHours
        - createdAt
      properties:
        id:
          type: integer
        jobId:
          type: integer
        name:
          type: string
        orderIndex:
          type: integer
        status:
          type: string
          enum: [not_started, in_progress, complete]
        estimatedHours:
          type: number
        actualHours:
          type: number
        createdAt:
          type: string
          format: date-time
...
    TimeEntry:
      type: object
      required:
        - id
        - employeeId
        - employeeName
        - jobId
        - jobNumber
        - jobName
        - stageId
        - stageName
        - clockIn
        - createdAt
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeTitle:
          type: ["string", "null"]
        jobId:
          type: integer
        jobNumber:
          type: string
        jobName:
          type: string
        stageId:
          type: integer
        stageName:
          type: string
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: ["string", "null"]
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    ClockOutInput:
      type: object
      required: [entryId]
      properties:
        entryId:
          type: integer
...
    BomPart:
      type: object
      required: [quantity]
      properties:
        id:
          type: integer
        partMark:
          type: [string, "null"]
        quantity:
          type: integer
          minimum: 1
        profileType:
          type: [string, "null"]
        profileSize:
          type: [string, "null"]
        grade:
          type: [string, "null"]
        lengthIn:
          type: [number, "null"]
        description:
          type: [string, "null"]
        heatNumber:
          type: [string, "null"]
    BomAssembly:
      type: object
      required: [mark, quantity, parts, onHold]
      properties:
        id:
          type: integer
        mark:
          type: string
        quantity:
          type: integer
          minimum: 1
        description:
          type: [string, "null"]
        finish:
          type: [string, "null"]
        processingPath:
          type: [string, "null"]
        currentStage:
          type: [string, "null"]
        onHold:
          type: boolean
        notes:
          type: [string, "null"]
        inspectedOn:
          type: [string, "null"]
        station:
          type: [string, "null"]
        inspector:
          type: [string, "null"]
        parts:
          type: array
          items:
            $ref: "#/components/schemas/BomPart"
...